refactor(ui): extract ProfilePage presentational sub-components
Move the props-only ProgressRing, StatCard, SectionCard and TagInput widgets
out of ProfilePage.tsx into a co-located ProfilePage.components module. These
have no dependency on the page's state, so the move is behavior-preserving.
